# api-spec.yaml
openapi: 3.0.2
info:
  title: FastAPI
  version: 0.1.0
paths:
...
  /v1/validation/credentials:
    get:
      summary: Validate Key
      description: |-
        HTTP endpoint used for credential (key + service_id) validation.

        Executes functions from the usecase layer to validate the keys.

        No authentication needed at this endpoint, as it will only be called internally from
        the K8s cluster.

        Returns 204 status if the credentials are valid return 401 if the credentials are not valid.

        Parameters:
        - apikey: A header parameter received by the endpoint
        - service: A header parameter received bt the endpoint
        - connection (AsyncConnection): An asynchronous database connection object (psycopg2).

        Returns:
        - None.

        Raises:
        - HTTPException (status_code 401): If the credentials (key + service ID) are not valid.
      operationId: validate_key_v1_validation_credentials_get
      parameters:
        - required: true
          schema:
            title: Apikey
            type: string
          name: apikey
          in: header
        - required: true
          schema:
            title: Service
            type: string
          name: service
          in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ValidationResponse:
      type: object
      properties:
        msg:
          type: string
...
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
...
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string